// aeons-study-circle-app/src/app/models/IVendorModel.ts
export interface VendorModel {
  id:number;
  name: string;
  phone: string;
  email: string;
  planId: number;
  domain: string;
  websiteUrl:string;
  uniqueKey: string;
  merchantKey: string;
  address: string;
  themeColor:string;
  logoImg:[];
  statusId:number;
  selected:boolean;
}